2-update

+++ src/meals/meals.service.ts
@@ -22,9 +22,11 @@ export class MealsService {
   //? application...
 
   async findForRestaurant(restaurant: string) {
-    const meals = await this.MealsModel.find({$and: [{restaurant}, {state: "Active"}]}).select(['-__v', '-createdAt', '-updatedAt', '-state', '-position', '-restaurant', '-tag', '-restaurantCategory']);
+    const meals = await this.MealsModel.find({$and: [{restaurant}, {state: "Active"}]}).select(['-__v', '-createdAt', '-updatedAt', '-state', '-position', '-restaurant', '-tag']);
     for(const single of meals){
       if(single?.image) single.image = this.awsService.getUrl(single.image);
+      if(single?.points) single.points = single.points * single.price;
+      if(single?.pointsBack) single.pointsBack = (single.pointsBack / 100) * single.price;
     }
     return meals;
   }
@@ -37,6 +39,19 @@ export class MealsService {
       if(single?.image) single.image = this.awsService.getUrl(single.image);
     }
     return {data: meals, pages: Math.ceil(total / limitEntity.limit)};
+  }
+
+  async findForCategory(category: string, orderBy?: string) {
+    if(!isValidObjectId(category)) throw new BadRequestException("There isn't meals with this category id");
+    let sort :any = {position: 1};
+    if(orderBy === 'highestPrice') { sort = { price: -1}}
+    else if (orderBy === 'lowestPrice') { sort = { price: 1}}
+    else if (orderBy === 'recently') { sort = { _id: -1}}
+    const meals = await this.MealsModel.find({$and: [{category}, {state: "Active"}]}).select(['-position','-state', '-__v', '-createdAt', '-updatedAt']).sort(sort);
+    for(const single of meals){
+      if(single?.image) single.image = this.awsService.getUrl(single.image);
+    }
+    return meals;
   }
 
   async findForTag(tag: string) {
@@ -53,6 +68,8 @@ export class MealsService {
     const meals = await this.MealsModel.find({$and: [{restaurantCategory}, {state: "Active"}]}).select(['-__v', '-createdAt', '-updatedAt', '-state', '-position', '-restaurant', '-tag', '-restaurantCategory']);
     for(const single of meals){
       if(single?.image) single.image = this.awsService.getUrl(single.image);
+      if(single?.points) single.points = single.points * single.price;
+      if(single?.pointsBack) single.pointsBack = (single.pointsBack / 100) * single.price;
     }
     return meals;
   }
@@ -86,7 +103,7 @@ export class MealsService {
 
   async findAll(limitEntity: LimitEntity) {
     const startIndex = limitEntity.page * limitEntity.limit;
-    const meals = await this.MealsModel.find().sort({_id: -1}).limit(limitEntity.limit).skip(startIndex);
+    const meals = await this.MealsModel.find().sort({position: 1}).limit(limitEntity.limit).skip(startIndex).populate("category");
     const total = await this.MealsModel.countDocuments();
     for(const single of meals){
       if(single?.image) single.image = this.awsService.getUrl(single.image);
@@ -94,8 +111,18 @@ export class MealsService {
     return {data: meals, pages: Math.ceil(total / limitEntity.limit)};
   }
 
+  async findAllForRestaurant(limitEntity: LimitEntity) {
+    const startIndex = limitEntity.page * limitEntity.limit;
+    const meals = await this.MealsModel.find({$and: [{restaurant: limitEntity.user}, {state: "Active"}]}).sort({_id: -1}).limit(limitEntity.limit).skip(startIndex);
+    const total = await this.MealsModel.countDocuments({$and: [{restaurant: limitEntity.user}, {state: "Active"}]});
+    for(const single of meals){
+      if(single?.image) single.image = this.awsService.getUrl(single.image);
+    }
+    return {data: meals, pages: Math.ceil(total / limitEntity.limit)};
+  }
+
   async findOne(id: string) {
-    const meal = await this.MealsModel.findById(id);
+    const meal = await this.MealsModel.findById(id).populate({path: "category", select: {_id: 1}});
     if(meal?.image) meal.image = this.awsService.getUrl(meal.image);
     return meal;
   }
@@ -124,7 +151,7 @@ export class MealsService {
   async remove(id: string) {
     const {image} = await this.MealsModel.findOne({_id: id}, {image: 1, _id: 0});
     await this.MealsModel.findByIdAndDelete(id);
-    this.awsService.removeImage(image);
+    if(image) this.awsService.removeImage(image);
     return "Success";
   }
